Refactor out the active ack logic to a class
